backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
sqlalchemy==2.0.30
camelot-py==0.11.0
opencv-python-headless
PyPDF2==3.0.1
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0
werkzeug==3.0.3
